Implimented Product insertion

+++ Database/Bumo Games Database.sql
@@ -8,8 +8,8 @@ Create Table Products(
 Productid INT NOT NULL PRIMARY KEY IDENTITY(1,1),
 Name varchar(50) NOT NULL,
 Description VARCHAR(255) NOT NULL,
-UnitCost float NOT NULL,
-DisplayCost float NOT NULL,
+UnitCost MONEY NOT NULL,
+DisplayCost MONEY NOT NULL,
 Image VARCHAR(255) NOT NULL,
 CategoryID INT,
 Hidden BIT  NOT NULL)
